+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 Unreleased
 
+-   The description for a model's object, and for an attribute's field and
+    argument, is set from their docstrings. :issue:`19`
+
 
 ## Version 1.0.0